requirements.txt
numpy
noise
pygame
PyOpenGL
PyOpenGL_accelerate